tool/pyproject.toml
[tool.poetry]
name = "tool"
version = "0.1.0"
description = ""
authors = ["matsuken-718 <matsunaga@lpixel.net>"]

[tool.poetry.dependencies]
python = "^3.10"
PyGithub = "^1.55"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
